requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2023.10
cron_descriptor==1.4.0
decorator==4.4.2
Django==4.2.7
django-admin-extra-buttons==1.5.8
django-celery-beat==2.5.0
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-filter==23.3
django-modeltranslation==0.18.11
django-timezone-field==6.1.0
djangorestframework==3.14.0
environ==1.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
import-export==0.3.1
kombu==5.3.4
moviepy==1.0.3
numpy==1.26.4
Pillow==10.1.0
proglog==0.1.10
prompt-toolkit==3.0.43
psycopg2==2.9.9
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
